kitti eval truncated int to float

+++ lib/datasets/voc_eval.py
@@ -20,7 +20,7 @@ def parse_rec(filename):
     obj_struct = {}
     obj_struct['name'] = obj.find('name').text
     obj_struct['pose'] = obj.find('pose').text
-    obj_struct['truncated'] = int(obj.find('truncated').text)
+    obj_struct['truncated'] = float(obj.find('truncated').text)
     obj_struct['difficult'] = int(obj.find('difficult').text)
     bbox = obj.find('bndbox')
     obj_struct['bbox'] = [int(bbox.find('xmin').text),
